Singled out the wait for ... to resolve function into TimeUtils module
